Added arguments to executable see pk_install.py -h for more information Made changes to git.py syntax for better fail handling of existing resources Made changes to handling of shell scripts, installer now recognises missing scripts instead of crashing the application.

+++ pk_install.py
@@ -18,6 +18,7 @@
 import os
 import sys
 import logging
+import argparse as ap
 import datetime as dt
 from subprocess import STDOUT, check_call
 from colorama import Fore, Style
@@ -46,15 +47,23 @@ class PKInstall:
         #
         # Set tools root directory
         if 'root' in args:
-            self.__ROOT_PATH__ = args['root'].strip('/')
+            self.__ROOT_PATH__ = args['root'].rstrip('/')
 
         self.__GIT__ = '{}/git/'.format(self.__ROOT_PATH__)
         self.__TOOLS__ = '{}/tools/'.format(self.__ROOT_PATH__)
 
+        self.__LOG_FILE__ = '{}/pk_install.log'.format(self.__ROOT_PATH__)
+
         #
         # Define Environment variables used in external scripts
         os.environ['PKIROOT'] = self.__ROOT_PATH__
-        os.environ['PKIUID'] = "20001"
+        os.environ['PKILOG'] = self.__LOG_FILE__
+
+        # Note, remember to change the password manually before using the system!
+        os.environ['PKIUSR'] = 'user'
+        os.environ['PKIPSW'] = 'Kali1234!'
+
+        os.environ['PKIUID'] = '20001'
         if 'uid' in args and isinstance(args['uid'], int):
             os.environ['PKIUID'] = args['uid']
 
@@ -75,7 +84,7 @@ class PKInstall:
         else:
             log_level = logging.INFO
 
-        logging.basicConfig(format='%(levelname)s: %(asctime)s: %(message)s', filename='{}/pk_install.log'.format(self.__ROOT_PATH__), level=log_level)
+        logging.basicConfig(format='%(levelname)s: %(asctime)s: %(message)s', filename=self.__LOG_FILE__, level=log_level)
         logging.info('Started running PK Installer at {}'.format(dt.datetime.now()))
         if 'skip' in args:
             if 'apt' in args['skip']:
@@ -101,21 +110,23 @@ class PKInstall:
             self.__ACCESS_RIGHTS__ = args['access_rights']
 
         if self.__DEBUG__:
-            msg = "\n\tROOT PATH: {},\n\t" \
-                  "ACCESS RIGHTS: {},\n\t" \
-                  "INSTALL SOURCES: \n\t\t" \
-                  "apt: {},\n\t\t" \
-                  "git: {},\n\t\t" \
-                  "wget: {},\n\t\t" \
-                  "scripts: {},\n\t\t" \
-                  "SIMULATOR MODE: {}\n\t\t" \
-                  "LOG LEVEL: {}".format(self.__ROOT_PATH__,
+            msg = '\n\tROOT PATH: {},\n\t' \
+                  'ACCESS RIGHTS: {},\n\t' \
+                  'INSTALL SOURCES: \n\t\t' \
+                  'apt: {},\n\t\t' \
+                  'git: {},\n\t\t' \
+                  'wget: {},\n\t\t' \
+                  'scripts: {},\n\t\t' \
+                  'SIMULATOR MODE: {}\n\t\t' \
+                  'UID: {}\n\t\t' \
+                  'LOG LEVEL: {}'.format(self.__ROOT_PATH__,
                                          self.__ACCESS_RIGHTS__,
                                          self.__SOURCES__['apt'],
                                          self.__SOURCES__['git'],
                                          self.__SOURCES__['wget'],
                                          self.__SOURCES__['scripts'],
                                          self.__SIMULATOR__,
+                                         args['uid'],
                                          log_level)
             logging.debug(msg)
 
@@ -125,9 +136,9 @@ class PKInstall:
         # Defaults to: ~/git/ and ~/tools/
         if self.__SIMULATOR__:
             print(self.MENU, Fore.LIGHTRED_EX + 'SIMULATOR Mode Enabled' + Style.RESET_ALL)
-            logging.info("SIMULATOR: Running in simulator mode! "
-                         "This will only dump commands which would have been run on the system. "
-                         "It is recommended that this mode is run with the DEBUG config flag set for the best output")
+            logging.info('SIMULATOR: Running in simulator mode! '
+                         'This will only dump commands which would have been run on the system. '
+                         'It is recommended that this mode is run with the DEBUG config flag set for the best output')
 
         print(self.MENU, Fore.LIGHTBLUE_EX + 'Building directory structure' + Style.RESET_ALL)
         logging.debug('Attempting to build the directory structure')
@@ -135,44 +146,44 @@ class PKInstall:
 
         if self.__SOURCES__['scripts']:
             print(self.MENU, Fore.LIGHTBLUE_EX + 'Running initial scripts' + Style.RESET_ALL)
-            logging.debug("Attempting to run initial scripts")
+            logging.debug('Attempting to run initial scripts')
             self.__scripts_install()
 
         if self.__SOURCES__['apt']:
             print(self.MENU, Fore.LIGHTBLUE_EX + 'Installing packages with APT' + Style.RESET_ALL)
-            logging.debug("Attempting to install packages with apt")
+            logging.debug('Attempting to install packages with apt')
             self.__apt_install()
 
         if self.__SOURCES__['git']:
             print(self.MENU, Fore.LIGHTBLUE_EX + 'Fetching packages with GIT' + Style.RESET_ALL)
-            logging.debug("Attempting to fetch packages with git")
+            logging.debug('Attempting to fetch packages with git')
             self.__git_install()
 
         if self.__SOURCES__['wget']:
             print(self.MENU, Fore.LIGHTBLUE_EX + 'Fetching packages with WGET' + Style.RESET_ALL)
-            logging.debug("Attempting to fetch packages with wget")
+            logging.debug('Attempting to fetch packages with wget')
             self.__wget_install()
 
         if self.__SOURCES__['scripts']:
             print(self.MENU, Fore.LIGHTBLUE_EX + 'Running final scripts' + Style.RESET_ALL)
-            logging.debug("Attempting to run final scripts")
+            logging.debug('Attempting to run final scripts')
             self.__scripts_install(False)
 
-        print(self.MENU, Fore.LIGHTBLUE_EX + 'FINISHED ' + Style.RESET_ALL + "🍻")
+        print(self.MENU, Fore.LIGHTBLUE_EX + 'FINISHED ' + Style.RESET_ALL + '🍻')
         logging.info('Installation finished at {}'.format(dt.datetime.now()))
         return
 
     def __build_dir_tree(self):
-        logging.debug("Entered self.__build_dir_tree()")
+        logging.debug('Entered self.__build_dir_tree()')
 
         if os.path.isdir(self.__ROOT_PATH__) and os.access(self.__ROOT_PATH__, os.W_OK):
             try:
-                logging.debug("Creating git directory: {}".format(self.__GIT__))
+                logging.debug('Creating git directory: {}'.format(self.__GIT__))
                 if not os.path.isdir(self.__GIT__):
                     if self.__SOURCES__['git']:
                         if self.__SIMULATOR__:
-                            msg = "SIMULATOR: creating directory: {} " \
-                                  "with the following rights {}".format(self.__GIT__, self.__ACCESS_RIGHTS__)
+                            msg = 'SIMULATOR: creating directory: {} ' \
+                                  'with the following rights {}'.format(self.__GIT__, self.__ACCESS_RIGHTS__)
                             logging.info(msg)
                         else:
                             print(self.OK ,'\tCreating directory: {}'.format(self.__GIT__))
@@ -192,12 +203,12 @@ class PKInstall:
                             print(self.WARN , '\tThe directory: {} already exist'.format(new_dir))
                             logging.warning('The directory {} already exists, using existing directory'.format(new_dir))
 
-                logging.debug("Creating tools directory: {}".format(self.__TOOLS__))
+                logging.debug('Creating tools directory: {}'.format(self.__TOOLS__))
                 if not os.path.isdir(self.__TOOLS__):
                     if self.__SOURCES__['wget']:
                         if self.__SIMULATOR__:
-                            msg = "SIMULATOR: creating directory: {} " \
-                                  "with the following rights {}".format(self.__TOOLS__, self.__ACCESS_RIGHTS__)
+                            msg = 'SIMULATOR: creating directory: {} ' \
+                                  'with the following rights {}'.format(self.__TOOLS__, self.__ACCESS_RIGHTS__)
                             logging.info(msg)
                         else:
                             print(self.OK, '\tCreating directory: {}'.format(self.__TOOLS__))
@@ -215,7 +226,7 @@ class PKInstall:
         logging.debug('Entered self.__apt_install()')
         if os.geteuid() != 0:
             print(self.FAIL, '\tInstalling software with apt should only be done by ' + Fore.LIGHTRED_EX + 'root' + Style.RESET_ALL + ', skipping step...')
-            logging.warning('Installing software with apt should only be done by root, skipping step...')
+            logging.error('Installing software with apt should only be done by root, skipping step...')
             return
 
         try:
@@ -250,49 +261,37 @@ class PKInstall:
                     os.chdir('{}/{}'.format(os.getcwd(), key))
 
                 for pkg in packages[key]:
-                    if key == 'tools':
-                        url = packages[key][pkg]['url']
-                        cfg = ''
-                        if 'config' in packages[key][pkg]:
-                            cfg = '{}/{}'.format(self.__ROOT_PATH__, packages[key][pkg]['config'])
+                    url = packages[key][pkg]['url']
+                    cfg = ''
+                    if 'config' in packages[key][pkg]:
+                        cfg = '{}/{}'.format(self.__ROOT_PATH__, packages[key][pkg]['config'])
 
-                        if not os.path.isdir('{}/{}'.format(os.getcwd(), pkg)) or self.__SIMULATOR__:
-                            print(self.OK, '\tCloning {}'.format(url))
-                            logging.debug('Installing {} with git'.format(url))
-                            try:
+                    if not os.path.isdir('{}/{}'.format(os.getcwd(), pkg)) or self.__SIMULATOR__:
+                        print(self.OK, '\tCloning {}'.format(url))
+                        logging.debug('Installing {} with git'.format(url))
+                        try:
+                            if self.__SIMULATOR__:
+                                logging.info('SIMULATOR: running command: git clone {} {}'.format(url, pkg))
+                            else:
+                                check_call(['git', 'clone', url, pkg], stdout=open(os.devnull, 'wb'), stderr=STDOUT)
+
+                            if cfg != '':
                                 if self.__SIMULATOR__:
-                                    logging.info('SIMULATOR: running command: git clone {} {}'.format(url, pkg))
+                                    logging.info('SIMULATOR: running command: sh {}'.format(cfg))
                                 else:
-                                    check_call(['git', 'clone', url, pkg], stdout=open(os.devnull, 'wb'), stderr=STDOUT)
-
-                                if cfg != '':
-                                    if self.__SIMULATOR__:
-                                        logging.info('SIMULATOR: running command: sh {}'.format(cfg))
-                                    else:
-                                        try:
-                                            print(self.OK, '\tConfiguring {}'.format(key))
-                                            logging.debug('Running {} config script'.format(pkg))
-                                            check_call(['sh', cfg], stdout=open(os.devnull, 'wb'), stderr=STDOUT)
-                                        except OSError as e:
-                                            print(self.FAIL, '\tFailed configuring {}'.format(key))
-                                            logging.warning(e)
-                            except OSError as e:
-                                print(self.FAIL, e)
-                                logging.warning(e)
-                        else:
-                            print(self.WARN, '\tThe directory {0}/{1} already exists, skipping tool {1}'.format(os.getcwd(), pkg))
-                            logging.warning('The directory {0}/{1} already exists, skipping tool {1}'.format(os.getcwd(), pkg))
+                                    try:
+                                        print(self.OK, '\tConfiguring {}'.format(key))
+                                        logging.debug('Running {} config script'.format(pkg))
+                                        check_call(['sh', cfg], stdout=open(os.devnull, 'wb'), stderr=STDOUT)
+                                    except OSError as e:
+                                        print(self.FAIL, '\tFailed configuring {}'.format(key))
+                                        logging.error(e)
+                        except OSError as e:
+                            print(self.FAIL, e)
+                            logging.error(e)
                     else:
-                        print(self.OK, '\tCloning {}'.format(pkg))
-                        logging.debug('Installing {} with git'.format(pkg))
-                        if self.__SIMULATOR__:
-                            logging.info('SIMULATOR: running command: git clone {}'.format(pkg))
-                        else:
-                            try:
-                                check_call(['git', 'clone', pkg], stdout=open(os.devnull, 'wb'), stderr=STDOUT)
-                            except OSError as e:
-                                print(self.FAIL, '\tFailed installing {}'.format(pkg))
-                                logging.warning(e)
+                        print(self.WARN, '\tThe directory {0}/{1} already exists, skipping resource {1}'.format(os.getcwd(), pkg))
+                        logging.warning('The directory {0}/{1} already exists, skipping resource {1}'.format(os.getcwd(), pkg))
 
                 if not self.__SIMULATOR__:
                     logging.debug('Changing directory to: {}'.format(self.__GIT__))
@@ -300,7 +299,7 @@ class PKInstall:
                     logging.debug('CWD: {}'.format(os.getcwd()))
         except OSError as e:
             print(self.FAIL, '\t' + e)
-            logging.warning(e)
+            logging.error(e)
 
         if not self.__SIMULATOR__:
             logging.debug('Changing directory to: {}'.format(self.__ROOT_PATH__))
@@ -326,7 +325,7 @@ class PKInstall:
                     check_call(['wget', pkg], stdout=open(os.devnull, 'wb'), stderr=STDOUT)
         except OSError as e:
             print(self.FAIL, '\t' + e)
-            logging.warning(e)
+            logging.error(e)
 
         if not self.__SIMULATOR__:
             logging.debug('Changing directory to: {}'.format(self.__ROOT_PATH__))
@@ -350,10 +349,18 @@ class PKInstall:
                     logging.info('SIMULATOR: running command: sh {}'.format(pkg))
                 else:
                     # TODO; Find a way to redirect stdout from scripts to logger
-                    check_call(['sh', pkg], stdout=open(os.devnull, 'wb'), stderr=STDOUT, env=dict(os.environ))
+                    try:
+                        if os.path.isfile(pkg):
+                            check_call(['sh', pkg], stdout=open(os.devnull, 'wb'), stderr=STDOUT, env=dict(os.environ))
+                        else:
+                            print(self.FAIL, '\t{}/{} does not exist'.format(self.__ROOT_PATH__, pkg))
+                            logging.error('{}/{} does not exist'.format(self.__ROOT_PATH__, pkg))
+                    except OSError as e:
+                        print(self.FAIL, '\t' + e)
+                        logging.error(e)
         except OSError as e:
             print(self.FAIL, '\t' + e)
-            logging.warning(e)
+            logging.error(e)
 
         logging.debug('Scripts done.')
         return
@@ -404,9 +411,58 @@ class PKInstall:
 .+´ Post Kali Installer v1.0 `+.
         '''
 
-if __name__ == '__main__':
-    config = {
 
-    }
-    pki = PKInstall(config)
-    pki.install()
+def main():
+    p = ap.ArgumentParser()
+
+    p.add_argument('-sA', '--skip-apt', default=False, action='store_true', help='Do not install APT packages')
+    p.add_argument('-sG', '--skip-git', default=False, action='store_true', help='Do not fetch GIT packages')
+    p.add_argument('-sS', '--skip-scripts', default=False, action='store_true', help='Do not run scripts')
+    p.add_argument('-sW', '--skip-wget', default=False, action='store_true', help='Do not fetch WGET packages')
+
+    p.add_argument('-D', '--debug', default=False, action='store_true', help='Run in debug mode')
+    p.add_argument('-S', '--simulate', default=False, action='store_true', help='Run in simulator mode')
+
+    p.add_argument('-R', '--set-root', type=str, help='Set install root directory, defaults to current working directory')
+    p.add_argument('-cU', '--set-config-uid', type=str, help='Set uid for use in config scripts')
+    #p.add_argument('-A', '--set-access-rights', type=str, help='Set access rights for created directories')
+
+    try:
+        a = p.parse_args()
+        config = _build_config(a)
+
+        pki = PKInstall(config)
+        pki.install()
+    except ValueError as ve:
+        sys.exit(ve)
+    return
+
+
+def _build_config(args):
+    config = {}
+    if args.skip_apt or args.skip_git or args.skip_wget or args.skip_scripts:
+        config['skip'] = {}
+        if args.skip_apt:
+            config['skip']['apt'] = args.skip_apt
+        if args.skip_git:
+            config['skip']['git'] = args.skip_git
+        if args.skip_wget:
+            config['skip']['wget'] = args.skip_wget
+        if args.skip_scripts:
+            config['skip']['scripts'] = args.skip_scripts
+    if args.set_root:
+        config['root'] = args.set_root
+    if args.debug:
+        config['DEBUG'] = args.debug
+    if args.simulate:
+        config['SIMULATOR'] = args.simulate
+    if args.set_config_uid:
+        config['uid'] = args.set_config_uid
+    # if args.set_access_rights:
+    #     config['access_rights'] = args.set_access_rights
+
+    return config
+
+
+if __name__ == '__main__':
+    main()
